implement JKS/PKCS12 format support add load function add new dependency 'github.com/corbym/gocrest' for testing add new dependency 'software.sslmate.com/src/go-pkcs12' for PKCS12 support
