+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/pavlo-v-chernykh/keystore-go/v4
 
 go 1.17
+
+require (
+	github.com/corbym/gocrest v1.0.6
+	software.sslmate.com/src/go-pkcs12 v0.5.0
+)
+
+require golang.org/x/crypto v0.11.0 // indirect
